+++ routes/tournaments.js
@@ -54,13 +54,15 @@ router.get('/admin', function (req, res, next) {
         axios.get(req.app.locals.kcapp.api + '/tournament/groups'),
         axios.get(req.app.locals.kcapp.api + '/player'),
         axios.get(req.app.locals.kcapp.api + '/office'),
+        axios.get(req.app.locals.kcapp.api + '/venue'),
         axios.get(req.app.locals.kcapp.api + '/match/modes'),
         axios.get(req.app.locals.kcapp.api + '/match/types'),
-    ]).then(axios.spread((groups, players, offices, modes, types) => {
+    ]).then(axios.spread((groups, players, offices, venues, modes, types) => {
         res.marko(tournamentsAdminTemplate, {
             groups: groups.data,
             players: players.data,
             offices: offices.data,
+            venues: venues.data,
             modes: modes.data,
             types: types.data
         });
@@ -174,6 +176,9 @@ router.post('/admin', function (req, res, next) {
                         starting_score: group.score
                     }],
                     tournament_id: tournament.id
+                }
+                if (req.body.venue_id !== -1) {
+                    matchBody.venue_id = req.body.venue_id;
                 }
                 createMatch(req, matchBody);
             }